+++ examples/main/shared/viewer/ViewerPlugins.jsx
@@ -1,4 +1,5 @@
 import theme from '../theme/theme';
+import { BUTTON_TYPE } from 'wix-rich-content-plugin-button/dist/module.viewer';
 import { VIDEO_TYPE, videoTypeMapper } from 'wix-rich-content-plugin-video/dist/module.viewer';
 import { dividerTypeMapper } from 'wix-rich-content-plugin-divider/dist/module.viewer';
 import { HTML_TYPE, htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer';
@@ -133,6 +134,9 @@ const config = {
     getVideoUrl: src => `https://video.wixstatic.com/${src.pathname}`,
   },
   uiSettings,
+  [BUTTON_TYPE]: {
+    // onClick: () => {},
+  },
 };
 
 export const getConfig = (additionalConfig = {}) => {