+++ packages/plugin-button/web/src/components/button-viewer.jsx
@@ -1,24 +1,28 @@
+/* eslint-disable jsx-a11y/no-static-element-interactions */
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/default-styles.scss';
 class ButtonViewer extends PureComponent {
   render() {
-    const { theme } = this.props;
+    const { theme, onClick } = this.props;
     this.styles = this.styles || mergeStyles({ styles, theme });
     const { url, style, target, rel, buttonText } = this.props;
+    const isActionButton = Boolean(onClick);
+    const Component = isActionButton ? 'div' : 'a';
+    let props = { className: this.styles.button_container, style };
+    props = isActionButton
+      ? { onClick, ...props }
+      : {
+          href: url,
+          target,
+          rel,
+          ...props,
+        };
     return (
-      <div>
-        <a
-          className={this.styles.button_container}
-          href={url}
-          style={style}
-          target={target}
-          rel={rel}
-        >
-          <div className={this.styles.button_text}>{buttonText}</div>
-        </a>
-      </div>
+      <Component {...props}>
+        <div className={this.styles.button_text}>{buttonText}</div>
+      </Component>
     );
   }
 }
@@ -30,6 +34,7 @@ ButtonViewer.propTypes = {
   rel: PropTypes.string,
   buttonText: PropTypes.string,
   theme: PropTypes.object.isRequired,
+  onClick: PropTypes.func,
 };
 
 export default ButtonViewer;