+++ CHANGELOG.md
@@ -15,10 +15,12 @@
     Click to see more.
   </summary>
   
+### :rocket: New Feature
+- `button`
+  - [#953](https://github.com/wix-incubator/rich-content/pull/953) action button: by passing onClick callback, on click triggers it
 ### :bug: Bug Fix
 - `editor`
-  - [#951](https://github.com/wix-incubator/rich-content/pull/951) fix: tooltip opacity
-  
+  - [#951](https://github.com/wix-incubator/rich-content/pull/951) fix: tooltip opacity 
 </details>
 <hr/>